fix: coolsms 환경변수 수정
